pyproject.toml
[project]
name = "hackupc2025-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "ferran",email = "0ferransogas4@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<3.14"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "openai-whisper @ git+https://github.com/openai/whisper.git",
    "torch (>=2.7.0,<3.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-google-genai (>=2.1.4,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "torchaudio (>=2.7.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
